improve robustness of yaml import
