+++ tests/components/flux_led/test_init.py
@@ -4,27 +4,9 @@ from __future__ import annotations
 from unittest.mock import patch
 
 from homeassistant.components import flux_led
-from homeassistant.components.flux_led.const import (
-    CONF_COLORS,
-    CONF_CUSTOM_EFFECT,
-    CONF_CUSTOM_EFFECT_COLORS,
-    CONF_CUSTOM_EFFECT_SPEED_PCT,
-    CONF_CUSTOM_EFFECT_TRANSITION,
-    CONF_DEVICES,
-    CONF_SPEED_PCT,
-    CONF_TRANSITION,
-    DOMAIN,
-)
-from homeassistant.components.light import DOMAIN as LIGHT_DOMAIN
+from homeassistant.components.flux_led.const import DOMAIN
 from homeassistant.config_entries import ConfigEntryState
-from homeassistant.const import (
-    CONF_HOST,
-    CONF_MODE,
-    CONF_NAME,
-    CONF_PLATFORM,
-    CONF_PROTOCOL,
-    EVENT_HOMEASSISTANT_STARTED,
-)
+from homeassistant.const import CONF_HOST, EVENT_HOMEASSISTANT_STARTED
 from homeassistant.core import HomeAssistant
 from homeassistant.setup import async_setup_component
 from homeassistant.util.dt import utcnow
@@ -74,51 +56,3 @@ async def test_config_entry_retry(hass: HomeAssistant) -> None:
         await async_setup_component(hass, flux_led.DOMAIN, {flux_led.DOMAIN: {}})
         await hass.async_block_till_done()
         assert config_entry.state == ConfigEntryState.SETUP_RETRY
-
-
-async def test_migrate_from_yaml(hass: HomeAssistant) -> None:
-    """Test migrate from yaml."""
-    config = {
-        LIGHT_DOMAIN: [
-            {
-                CONF_PLATFORM: DOMAIN,
-                CONF_DEVICES: {
-                    IP_ADDRESS: {
-                        CONF_NAME: "flux_lamppost",
-                        CONF_PROTOCOL: "ledenet",
-                        CONF_CUSTOM_EFFECT: {
-                            CONF_SPEED_PCT: 30,
-                            CONF_TRANSITION: "strobe",
-                            CONF_COLORS: [[255, 0, 0], [255, 255, 0], [0, 255, 0]],
-                        },
-                    }
-                },
-            },
-            {
-                CONF_PLATFORM: "anything else",
-            },
-        ],
-        "garbage": [{}],
-    }
-    with _patch_discovery(), _patch_wifibulb():
-        await async_setup_component(hass, DOMAIN, config)
-        await hass.async_block_till_done()
-
-    migrated_entry = None
-    for entry in hass.config_entries.async_entries(DOMAIN):
-        if entry.unique_id == MAC_ADDRESS:
-            migrated_entry = entry
-            break
-
-    assert migrated_entry is not None
-    assert migrated_entry.data == {
-        CONF_HOST: IP_ADDRESS,
-        CONF_NAME: "flux_lamppost",
-        CONF_PROTOCOL: "ledenet",
-    }
-    assert migrated_entry.options == {
-        CONF_MODE: "auto",
-        CONF_CUSTOM_EFFECT_COLORS: "[[255, 0, 0], [255, 255, 0], [0, 255, 0]]",
-        CONF_CUSTOM_EFFECT_SPEED_PCT: 30,
-        CONF_CUSTOM_EFFECT_TRANSITION: "strobe",
-    }
